chrome extension mega commit 🙈
This embarassingly large commit adds most of what is needed to render
and build the chrome extension version of this app. It adds some new
components to support a chrome extension structure. It also modifies
some existing components to be agnostic to the rendering context -- mostly
this means whether the component has a next router or not, which does not
exist in the chrome extension.